[revert] a8c7fa2([refactor] modal 관련 로직 core로 이동) commit revert
